separated tests into multiple files
