+++ kassaSystem/kassaSystem/kassaSystem.csproj
@@ -53,7 +53,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Tests\KassaSystemTests.cs" />
+    <Compile Include="Tests\TestsList.cs" />
     <Compile Include="kassaSystem.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -62,6 +62,9 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tests\TestSetup.cs" />
+    <Compile Include="Tests\TestsButtons.cs" />
+    <Compile Include="Tests\TestsText.cs" />
     <EmbeddedResource Include="kassaSystem.resx">
       <DependentUpon>kassaSystem.cs</DependentUpon>
     </EmbeddedResource>
